+++ test/test_hydraulic.py
@@ -35,46 +35,6 @@ def test_cavitation_factor_increases_as_segment_water_potential_decreases():
         assert all(x >= y for x, y in zip(cavitation, cavitation[1:]))
 
 
-def test_def_param_soil_returns_the_right_soil_property_values():
-    ref_values = {'Sand': (0.045, 0.430, 0.145, 2.68, 712.8),
-                  'Loamy_Sand': (0.057, 0.410, 0.124, 2.28, 350.2),
-                  'Sandy_Loam': (0.065, 0.410, 0.075, 1.89, 106.1),
-                  'Loam': (0.078, 0.430, 0.036, 1.56, 24.96),
-                  'Silt': (0.034, 0.460, 0.016, 1.37, 6.00),
-                  'Silty_Loam': (0.067, 0.450, 0.020, 1.41, 10.80),
-                  'Sandy_Clay_Loam': (0.100, 0.390, 0.059, 1.48, 31.44),
-                  'Clay_Loam': (0.095, 0.410, 0.019, 1.31, 6.24),
-                  'Silty_Clay_Loam': (0.089, 0.430, 0.010, 1.23, 1.68),
-                  'Sandy_Clay': (0.100, 0.380, 0.027, 1.23, 2.88),
-                  'Silty_Clay': (0.070, 0.360, 0.005, 1.09, 0.48),
-                  'Clay': (0.068, 0.380, 0.008, 1.09, 4.80)}
-    module_values = hydraulic.def_param_soil()
-    for soil_class, soil_properties in ref_values.items():
-        assert all(x == y for x, y in zip(soil_properties, module_values[soil_class]))
-
-
-def test_def_param_soil_returns_Custom_soil_properties_when_provided():
-    custom_soil = (0.02, 0.3, 0.03, 1.5, 25)
-    assert hydraulic.def_param_soil(custom_soil)['Custom'] == custom_soil
-
-
-def test_soil_water_potential_decreases_as_water_withdrawal_increases():
-    for soil_class in hydraulic.def_param_soil().keys():
-        psi_soil = [hydraulic.soil_water_potential(psi_soil_init=0., water_withdrawal=w, soil_class=soil_class,
-                                                   soil_total_volume=1, psi_min=-3.)
-                    for w in arange(0, 3, 0.1)]
-        assert all(x >= y for x, y in zip(psi_soil, psi_soil[1:]))
-
-
-def test_soil_water_potential_drops_faster_for_small_soil_reservoirs_than_bigger_ones():
-    for soil_class in hydraulic.def_param_soil().keys():
-        psi_soil_small = hydraulic.soil_water_potential(psi_soil_init=0., water_withdrawal=1., soil_class=soil_class,
-                                                        soil_total_volume=1., psi_min=-3.)
-        psi_soil_big = hydraulic.soil_water_potential(psi_soil_init=0., water_withdrawal=1., soil_class=soil_class,
-                                                      soil_total_volume=2., psi_min=-3.)
-        assert psi_soil_small < psi_soil_big
-
-
 def test_hydraulic_prop_attributes_the_right_properties_to_the_different_organs():
     simple_shoot = potted_syrah()
     simple_shoot.node(simple_shoot.root).vid_base = architecture.get_mtg_base(simple_shoot, vtx_label='inT')
